fix(mobile-client): fix e2e tests - reorganize components in mobile-client in their own folder - fix cypress tests for mobile client - add storybook actions config for todo item mobile client component
