+++ templates/mantis-todo/apps/mobile-client-e2e/src/e2e/app.cy.ts
@@ -1,4 +1,4 @@
-import { getGreeting } from '../support/app.po';
+import { getTitle } from '../support/app.po';
 
 describe('mobile-client-e2e', () => {
   beforeEach(() => cy.visit('/'));
@@ -8,6 +8,6 @@ describe('mobile-client-e2e', () => {
     cy.login('my-email@something.com', 'myPassword');
 
     // Function helper example, see `../support/app.po.ts` file
-    getGreeting().contains(/Welcome/);
+    getTitle().contains(/Todos/);
   });
 });
