+++ templates/mantis-todo/apps/mobile-client-e2e/src/support/app.po.ts
@@ -1,1 +1,1 @@
-export const getGreeting = () => cy.get('h1');
+export const getTitle = () => cy.get('ion-title');
