+++ templates/mantis-todo/apps/mobile-client/src/app/components/todo-item/todo-item.component.ts
